docs: Input 컴포넌트 마이그레이션
