+++ ui-kit/src/stories/Components/Input/data.ts
@@ -1,0 +1,13 @@
+import { TextInputType } from 'src/components/Input';
+
+export const types: TextInputType[] = [
+  'text',
+  'email',
+  'number',
+  'password',
+  'search',
+  'tel',
+  'url',
+];
+export const covertToTitlecase = (s: string) =>
+  `${s.charAt(0).toUpperCase()}${s.slice(1, s.length)}`;
